requirements.txt
astropy
bokeh
george
matplotlib
numpy>=1.16.4
requests
scipy
sncosmo
tqdm
iminuit
sndata>=0.6.0
pyyaml
sfdmap

# For documentation building
sphinx
sphinx-argparse
